[FF-41] - Moved articles into a Modal instead of new path
