+++ src/App.css
@@ -233,3 +233,80 @@ html,
 .article-p {
   color: red; /* Just a toy example of how we can control the article's specifically */
 }
+
+/* Modal container */
+.modal {
+  display: flex;
+  position: fixed;
+  z-index: 1001; /* Ensure it is above other content */
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0, 0, 0, 0.5); /* Semi-transparent background */
+  justify-content: center;
+  align-items: center;
+}
+
+/* Modal content */
+.modal-content {
+  background-color: #fff;
+  padding: 20px;
+  border: 1px solid #888;
+  width: 80%; /* Adjust width as needed */
+  max-width: 600px; /* Ensure modal doesn't get too wide */
+  border-radius: 10px;
+  position: relative;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Add shadow for better visibility */
+  overflow-y: auto; /* Allow scrolling within the modal if content is too long */
+  max-height: 80vh; /* Ensure modal doesn't exceed viewport height */
+}
+
+/* Close button */
+.close {
+  color: #aaa;
+  float: right;
+  font-size: 28px;
+  font-weight: bold;
+  position: absolute;
+  right: 10px;
+  top: 10px;
+}
+
+.close:hover,
+.close:focus {
+  color: black;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+/* Optional: Add transition for modal appearance */
+.modal-content {
+  animation: fadeIn 0.3s;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+    transform: scale(0.9);
+  }
+  to {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
+
+/* Additional styling for the modal if needed */
+.article-modal h1 {
+  margin-top: 0;
+}
+
+.article-modal p {
+  margin-bottom: 1rem;
+}
+
+.article-modal img {
+  max-width: 100%;
+  height: auto;
+}
